Fix Roadie for Rails 5
This is a temporary solution to get past rails/rails@7b122d3 until I or
someone else gets around to upgrading roadie-rails for Rails 5.

In the rails commit rails/rails@7b122d3 the method `collect_responses`
was made private. This means that when `method_defined?` is called
`collect_responses` is not defined because it's private. This causes the
conditional to fall through to the else and look for
`collect_responses_and_parts_order` which was removed from Rails some
time ago.

The commit here should not be used past the temporary point of getting
roadie-rails working with Rails 5.

I wanted to monkey-patch this but decided to fork instead for 2 reasons:
1) It's obvious something is forked becasue we generally look at the
Gemfile first, not in core_ext or in initializers.

2) I want this to be temporary and feel temporary. Monkey patches are
easier to be ok with in your codebase.