Fix alias_method_chain warning
While waiting for the Rails 5 branch to be finished up on roadie-rails
this will fix the deprecation warning for `alias_method_chain` by
replacing it with `prepend`. This required refactoring existing methods
because now the correct way to do this is to keep the method names and
then call `super`.